+++ src/parsing/tokenization/tokenization_utils.c
@@ -14,8 +14,8 @@
 
 int	check_quote(char *line)
 {
-	int		i;
-	int		quote;
+	size_t	i;
+	size_t	quote;
 	char	quotetype;
 
 	i = 0;
@@ -39,7 +39,7 @@ int	check_quote(char *line)
 	return (0);
 }
 
-static void	syntax_error(char *line, int *i)
+static void	syntax_error(char *line, size_t *i)
 {
 	if (line[*i] == '\0')
 		print_error(0, "syntax error near unexpected token 'newline'");
@@ -51,7 +51,7 @@ static void	syntax_error(char *line, int *i)
 	}
 }
 
-char	**get_data(t_input **input, char *line, int *i)
+char	**get_data(t_input **input, char *line, size_t *i)
 {
 	char	**data;
